Change release artifacts location and use assembly attributes generation
Rework approach so that for Release builds artifacts are put to the
"ReleaseArtifacts" project in the root. That is to better support
multi-target frameworks where there are many folders present.
The old approach with hardcoded paths will blow up here.

Also switch to Assembly attributes generation. This approach allows to
pass assembly version as build parameter without need to physically patch
the AssemblyInfo.cs files.